src/pyproject.toml
[tool.poetry]
name = "federated-engineers-portal-backend"
version = "0.1.9"
description = "A login portal for federated engineers to do their work"
authors = ["Ben van der Harg <ben.j.van.der.harg@postnl.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
toml = "^0.10.2"
uvicorn = "^0.29.0"
dependency-injector-fork = "^4.42.1"
boto3 = "^1.34.110"
dogpile-cache = "^1.3.3"
redis = "^5.0.4"
requests = "^2.32.2"
jinja2 = "^3.1.4"
starsessions = "^2.1.3"
cryptography = "^42.0.7"
mypy-boto3-cognito-identity = "^1.34.0"
mypy-boto3-cognito-idp = "^1.34.101"
mypy-boto3-secretsmanager = "^1.34.109"
mypy-boto3-sts = "^1.34.0"
boto3-stubs = "^1.34.109"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.9.0"
types-requests = "^2.31.0.20240406"
types-toml = "^0.10.8.20240310"
types-python-jose = "^3.3.4.20240106"
pytest = "^8.3.2"

[tool.mypy]
explicit_package_bases = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
